Refactored Viper Pilot Database UI
Moved edit pilot database control from DLNK system page to aux control so the database can be edited outside the DLNK page. This feels a little cleaner.
